annotated prop path checkpoint at 9/29/24 3:05P

+++ src/utils/go/evaluator/eval2_annotate_clone_counts.go
@@ -2,7 +2,7 @@ package evaluator
 
 type CloneCountResult struct {
 	PropertyPath *CloneCountPropertyPath
-	Count        int
+	Count        Float
 }
 
 // Annotates property paths with clone counts.
